Re-factored command and handlers
There now are layers of handler queues. The migration CLI command
creates a composite Migration handler, which will run Migration Module
handlers - one for every module. The Relationships module handler uses
the Relationships migrator to process a batch of Relationship records.

Each layer re-uses a Progress specific to that layer, which visually
displays the progress of that layer.